+++ unique_toolkit/unique_toolkit/components/internal_search/__init__.py
@@ -1,0 +1,45 @@
+from .base import (
+    DEFAULT_LIMIT_CHUNK_RELEVANCY_SORT_DISABLED,
+    DEFAULT_LIMIT_CHUNK_RELEVANCY_SORT_ENABLED,
+    HasChunkRelevancySorter,
+    InternalSearchBaseService,
+    InternalSearchConfig,
+    InternalSearchProgressMessage,
+    InternalSearchResult,
+    InternalSearchStage,
+    InternalSearchState,
+    SearchStringResult,
+    TInternalSearchDeps,
+)
+from .chat import ChatInternalSearchDeps, ChatInternalSearchService
+from .knowledge_base import (
+    UNSET,
+    KnowledgeBaseInternalSearchConfig,
+    KnowledgeBaseInternalSearchDeps,
+    KnowledgeBaseInternalSearchService,
+    KnowledgeBaseInternalSearchState,
+)
+
+__all__ = [
+    # base
+    "HasChunkRelevancySorter",
+    "InternalSearchConfig",
+    "InternalSearchBaseService",
+    "InternalSearchProgressMessage",
+    "InternalSearchResult",
+    "InternalSearchStage",
+    "InternalSearchState",
+    "SearchStringResult",
+    "TInternalSearchDeps",
+    "DEFAULT_LIMIT_CHUNK_RELEVANCY_SORT_ENABLED",
+    "DEFAULT_LIMIT_CHUNK_RELEVANCY_SORT_DISABLED",
+    # chat
+    "ChatInternalSearchDeps",
+    "ChatInternalSearchService",
+    # knowledge base
+    "KnowledgeBaseInternalSearchConfig",
+    "KnowledgeBaseInternalSearchDeps",
+    "KnowledgeBaseInternalSearchService",
+    "KnowledgeBaseInternalSearchState",
+    "UNSET",
+]